requirements.txt
Django>=2.2,<2.3
wagtail>=2.6,<2.7
django-environ==0.4.5
django-extensions==2.2.5
django-webpack-loader==0.6.0
gunicorn==19.9.0
djangorestframework==3.10.3
